+++ CHANGES.md
@@ -4,3 +4,4 @@
 
 - Switch from `requests` to `httpx` library.
 - Add `--verify_size` switch to disable file size sanity check.
+- Add auto-retry with backoff on intermittent server errors